agent/requirements.txt
pynput==1.7.6
requests==2.31.0
pyyaml==6.0.1
